+++ Branch-SDK/Branch-SDK/Branch.h
@@ -125,7 +125,6 @@ extern NSString * const BNCCanonicalIdList;
 extern NSString * const BNCPurchaseAmount;
 extern NSString * const BNCPurchaseCurrency;
 extern NSString * const BNCCanonicalIdList;
-extern NSString * const BNCPurchaseAmount;
 extern NSString * const BNCRegisterViewEvent;
 extern NSString * const BNCAddToWishlistEvent;
 extern NSString * const BNCAddToCartEvent;