+++ Branch-TestBed-Swift/TestBed-Swift/AppDelegate.swift
@@ -91,7 +91,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, AdjustDelegate, AppsFlyer
                     
                     if clickedBranchLink {
                         let nc = self.window!.rootViewController as! UINavigationController
-                        let storyboard = UIStoryboard(name: "ContentView", bundle: nil)
+                        let storyboard = UIStoryboard(name: "Content", bundle: nil)
                         let contentViewController = storyboard.instantiateViewController(withIdentifier: "Content") as! ContentViewController
                         nc.pushViewController(contentViewController, animated: true)
                         
@@ -183,6 +183,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, AdjustDelegate, AppsFlyer
         if IntegratedSDKsData.activeTuneEnabled()! {
             Tune.measureSession()
         }
+
+        // AppsFlyer
+        AppsFlyerTracker.shared().trackAppLaunch()
+        // Your code here...
     }
     
     func applicationWillResignActive(_ application: UIApplication) {
@@ -296,6 +300,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate, AdjustDelegate, AppsFlyer
         
         AppsFlyerTracker.shared().appsFlyerDevKey = key
         AppsFlyerTracker.shared().appleAppID = "1160975066"
+        AppsFlyerTracker.shared().delegate = self
+        AppsFlyerTracker.shared().isDebug = true
     }
     
     func activateGoogleAnalytics() {