+++ Branch.framework/Versions/A/Headers/BranchDeepLinkingController.h
@@ -6,12 +6,19 @@
 //  Copyright (c) 2015 Branch Metrics. All rights reserved.
 //
 
+#if __has_feature(modules)
 @import Foundation;
 @import UIKit;
+#else
+#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+#endif
 
 @protocol BranchDeepLinkingControllerCompletionDelegate <NSObject>
 
-- (void)deepLinkingControllerCompleted __attribute__((deprecated(("This API is deprecated. Instead, use deepLinkingControllerCompletedFrom: viewController"))));;
+- (void)deepLinkingControllerCompleted
+    __attribute__((deprecated(("This API is deprecated. Instead, use deepLinkingControllerCompletedFrom: viewController"))));;
+
 - (void)deepLinkingControllerCompletedFrom:(UIViewController*) viewController;
 
 @end
@@ -21,6 +28,8 @@ typedef NS_ENUM(NSInteger, BNCViewControllerPresentationOption) {
     BNCViewControllerOptionPush,
     BNCViewControllerOptionPresent
 };
+
+#pragma mark - BranchDeepLinkingController Protocol
 
 @protocol BranchDeepLinkingController <NSObject>